db: ScheduleContainer: Change `key` type from integer to string  * Also add a database migration test for the newest version.

+++ app/src/main/kotlin/by/ntnk/msluschedule/network/api/original/NetworkHelper.kt
@@ -60,14 +60,14 @@ class NetworkHelper @Inject constructor(private val currentDate: CurrentDate) {
             createWeekRequestDataInstance(weekKey),
             createFacultyRequestDataInstance(studyGroup.faculty),
             createCourseRequestDataInstance(studyGroup.course),
-            createStudyGroupRequestDataInstance(studyGroup.key)
+            createStudyGroupRequestDataInstance(studyGroup.key.toInt())
         )
 
     fun getTeacherRequestDataList(teacher: Teacher, weekKey: Int): List<RequestData> =
         listOf(
             createYearRequestDataInstance(teacher.year),
             createWeekRequestDataInstance(weekKey),
-            createTeacherRequestDataInstance(teacher.key)
+            createTeacherRequestDataInstance(teacher.key.toInt())
         )
 
     fun parseDataFromHtmlBody(requestInfo: RequestInfo, response: String): Single<ScheduleFilter> {
